requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
python-docx
openpyxl
python-pptx
